perf(3B): switch APG guidance from F64 to F32 precision
Python reference uses .double() but empirical A/B testing confirms F32
produces audibly identical results. Waveform differs slightly at the
numerical level but is perceptually indistinguishable.

- APGMomentumBuffer: double -> float running_average
- APGWorkspace: all double vectors -> float
- All intermediate math: double -> float
- Epsilon thresholds: 1e-60 -> 1e-30 (float range)
- Halves APG memory bandwidth (~1.5MB -> ~768KB workspace)

+++ engine/src/hot-step-sampler.h
@@ -256,7 +256,7 @@ static int dit_ggml_generate(DiTGGML *           model,
     if (do_cfg) {
         apg_mbufs.reserve(N);
         for (int i = 0; i < N; i++) {
-            apg_mbufs.emplace_back((double) -apg_momentum);  // configurable momentum
+            apg_mbufs.emplace_back(-apg_momentum);  // configurable momentum
         }
         fprintf(stderr, "[DiT] CFG enabled: guidance_scale=%.1f, %s, N_graph=%d",
                 guidance_scale, batch_cfg ? "batched" : "2-pass", N_graph);
